Break down and split Types into multiple modules
Problem: It's difficult to navigate around Types, since it's too big,
and compiling it after changing something takes time.

Solution: Break it into mutiple modules.
